agregando validaciones

+++ api/controllers/category.js
@@ -54,26 +54,30 @@ function updateCategoryByID(req, res, next){
     const id = req.params.categoryID;
     const cambios = req.body;
     const updateOps = {};
-    const q = category.where({_id: id });
+    const q = category.where({_id: id});
 
     for(const ops of req.body){
         updateOps[ops.propName] = ops.value;
     }
 
-    category.where({'_id':id}).update( {$set: updateOps}).exec()
-            .then(result =>{
-                if(result){
-                res.status(200).json(result)
-                }
-                else {
-                    console.log("en el else")
-                    res.status(404).json(result)
-                }
-            })
-            .catch(err =>{
-                res.status(500).json({error: err})
-            })
-
+     if(id.length===24){
+        category.where({'_id':id}).update( {$set: updateOps}).exec()
+                .then(result =>{
+                    if(result.nModified===1){
+                        res.status(200).json({message: "modificado con exito"})
+                    }
+                    else {                        
+                        res.status(404).json({message: 'no encontrado'})
+                    }
+                 })
+                .catch(err =>{
+                    res.status(500).json({error: err})
+                })
+        }
+        else{
+            res.status(404).json({message: 'error de id, incorrecto'})
+        }
+  
 }
 
 //eliminar una categoria
@@ -81,18 +85,24 @@ function deleteCategoryByID(req, res, next){
     console.log("dentro de la funcion delete category")
     const id = req.params.categoryID;
     
-    category.findByIdAndRemove(id).exec().then(result=>{
-        if(result){
-            res.status(200).json(result)
-        }
-        else {
-            res.status(404).json(result)
-        }
-    })
-    .catch(err =>{
-        console.log(err);
-        res.status(500).json({error: err})
-    })
+    if(id.length===24)
+    {
+        category.findByIdAndRemove(id).exec().then(result=>{
+            if(result){
+                res.status(200).json({message: "borrado con exito"})
+            }
+            else {
+                res.status(404).json({message: "error, ID no encontrado"})
+            }
+        })
+        .catch(err =>{
+            console.log(err);
+            res.status(500).json({error: err})
+        })
+    }
+    else{
+        res.status(404).json({message: "error ID incorrecto"}) 
+    }
    
 }
 
@@ -101,16 +111,18 @@ function getCategoryID(req,res,next) {
     console.log("aqui en el getcategory por ID");
     const id = req.params.categoryID;
     
+    if(id.length===24)
+    {
     category.findById(id)
             .exec()
             .then(categoryByID =>{
                 
                 if(categoryByID){
-                    console.log(categoryByID);
+                   // console.log(categoryByID);
                     res.status(200).json(categoryByID)
                 }
                 else {
-                    res.status(404).json({message: 'no valid entry for provided ID'})
+                    res.status(404).json({message: 'no encontrado, ID incorrecto'})
                 }
                 
             })
@@ -118,5 +130,9 @@ function getCategoryID(req,res,next) {
                 console.log(err);
                 res.status(500).json({error: err})
             })
+    }
+    else{
+        res.status(404).json({message: "no valid entry for provided ID"})
+    }
 }
 module.exports ={getCategories, newCategory, getCategoryID, deleteCategoryByID, updateCategoryByID};
